FEAT: update domain model
Change the GetByID method to take a Long instead of Object since the PK of Movie is a Long.

Dropped DB, removed migrations, and recreated migrations and DB to reflect the above changes.

+++ Migrations/RevisitedDbContextModelSnapshot.cs
@@ -18,7 +18,7 @@ namespace MovieCollectionRevisited.Migrations
 
             modelBuilder.Entity("MovieCollectionRevisited.Models.Movie", b =>
                 {
-                    b.Property<long>("MovieID")
+                    b.Property<int>("MovieID")
                         .ValueGeneratedOnAdd()
                         .HasColumnType("INTEGER");
